restricted naming of networks #4
