+++ src/main/java/net/cryptic_game/microservice/network/model/Member.java
@@ -62,15 +62,6 @@ public class Member extends Model {
 		} catch (SQLException e) {
 		}
 
-		ResultSet rs2 = db.getResult("SELECT `uuid` FROM `network` WHERE `owner`=?", device.toString());
-
-		try {
-			while (rs.next()) {
-				list.add(Network.get(UUID.fromString(rs.getString("uuid"))));
-			}
-		} catch (SQLException e) {
-		}
-
 		return list;
 	}
